Refactored Cart APIs
